t2-2 tehty

+++ t2-2.py
@@ -1,3 +1,3 @@
 import math
-sade = int(input("Ympyrän säde: "))
+sade = float(input("Ympyrän säde: "))
 print(math.pi *  (pow(sade, 2)))
